refactor : Move responsibiliy for matching
delete LottoMatch, LottoMatchTest
move responsibiliy matching lotto and winningLotto to Lottos

Please enter the commit message for your changes. Lines starting
